Refactor cards into new module
