--- eggsplode/views/__init__.py
@@ -1,8 +1,0 @@
-"""
-Contains all the views for the application.
-"""
-
-from .action import *
-from .base import *
-from .short import *
-from .starter import *
