+++ eggsplode/views/base.py
@@ -3,7 +3,7 @@ Contains the BaseView class which is used to create a Discord UI view for the ga
 """
 
 import discord
-from ..game_logic import ActionContext
+from ..ctx import ActionContext
 
 
 class BaseView(discord.ui.View):
